# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for managing data on oncological markers.",
        "title": "OncomarkerAPI",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/patient-diseases": {
            "get": {
                "description": "Retrieves a list of patient diseases.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PatientDisease"
                ],
                "summary": "Get patient disease list",
                "responses": {
                    "200": {
                        "description": "Patient disease list",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/model.PatientDisease"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Updates an existing patient disease.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PatientDisease"
                ],
                "summary": "Update patient disease",
                "parameters": [
                    {
                        "description": "Patient disease data",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.PatientDisease"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated patient disease data",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a new patient disease.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PatientDisease"
                ],
                "summary": "Create patient disease",
                "parameters": [
                    {
                        "description": "Patient disease data",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.PatientDisease"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Created patient disease data",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/patient-diseases/disease/{disease_id}": {
            "get": {
                "description": "Retrieves a list of patient diseases by disease ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PatientDisease"
                ],
                "summary": "Get patient disease list by disease",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Disease ID",
                        "name": "disease_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Patient disease list",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/model.PatientDisease"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/patient-diseases/patient/{patient_id}": {
            "get": {
                "description": "Retrieves a list of patient diseases by patient ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PatientDisease"
                ],
                "summary": "Get patient disease list by patient",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Patient ID",
                        "name": "patient_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Patient disease list",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/model.PatientDisease"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/patient-diseases/{patient_id}/{disease_id}": {
            "get": {
                "description": "Retrieves a patient disease by patient and disease ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PatientDisease"
                ],
                "summary": "Get patient disease by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Patient ID",
                        "name": "patient_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Disease ID",
                        "name": "disease_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Patient disease data",
                        "schema": {
                            "$ref": "#/definitions/model.PatientDisease"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes a patient disease by patient and disease ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PatientDisease"
                ],
                "summary": "Delete patient disease",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Patient ID",
                        "name": "patient_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Disease ID",
                        "name": "disease_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deleted patient disease data",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "handler.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "model.PatientDisease": {
            "type": "object",
            "properties": {
                "diagnosis": {
                    "type": "string"
                },
                "disease": {
                    "type": "string"
                },
                "patient": {
                    "type": "integer"
                },
                "stage": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}